src/requirements.txt
Django==3.2
django-behaviors
django-environ
whitenoise

djangorestframework==3.12.4
django-healthchecks

celery==4.4.7  # because of broken flower integration in 5.0.0
flower==0.9.5
requests==2.25.0
python-dateutil==2.8.1
gunicorn==20.0.4
uuid==1.30
datetime==4.3
pyyaml
psycopg2-binary==2.8.6
lxml==4.6.3

python-telegram-bot==13.5

fatsecret==0.2.3

redis
pandas==1.2.4

pytest-django==4.1.0
pytest-deadfixtures
pytest-env
pytest-freezegun
pytest-mock==3.4.0
pytest-randomly
factory-boy==3.2.0

requests_mock
freezegun
mixer

black